# apps/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ConvoLens API
  description: |
    API for the ConvoLens platform - WhatsApp chat summarization service.

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    - Default: 100 requests per 15 minutes
    - Auth endpoints: 5 attempts per 15 minutes

    ## Correlation IDs
    Include `x-correlation-id` header for request tracing.

  version: 1.0.0
  contact:
    name: ConvoLens Support
    email: support@convolens.neuralliquid.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.convolens.neuralliquid.ai
    description: Production server
...
paths:
...
  /api/chat-export:
    get:
      summary: List stored conversation intakes
      description: List up to 100 durable conversation records for the signed-in user
      tags:
        - Chat Export
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User-scoped conversation list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      conversations:
                        type: array
                        items:
                          $ref: '#/components/schemas/ConversationSummary'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/chat-export/{id}:
    get:
      summary: Get a stored conversation intake
      description: Load one durable conversation and its messages for the signed-in user
      tags:
        - Chat Export
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User-scoped conversation detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      conversation:
                        $ref: '#/components/schemas/ConversationDetail'
        '404':
          description: Conversation not found for this user
    delete:
      summary: Delete a stored conversation intake
      description: Permanently delete one conversation and its messages for the signed-in owner
      tags:
        - Chat Export
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Conversation deleted
        '401':
          description: Not authenticated
        '404':
          description: Conversation not found for this user
  /api/chat-export/upload:
    post:
      summary: Upload chat export file
      description: Upload a WhatsApp chat export text file for processing
      tags:
        - Chat Export
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: WhatsApp chat export .txt file
              required:
                - file
      responses:
        '200':
          description: File processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatExportResponse'
        '400':
          description: Invalid file format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/chat-export/{id}/summary:
    get:
      summary: Get a conversation catch-up
      description: Load the persisted, evidence-linked catch-up for an owned conversation
      tags:
        - Chat Export
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Catch-up result, or null when none has been generated
        '401':
          description: Not authenticated
        '404':
          description: Conversation not found for this user
    post:
      summary: Generate a conversation catch-up
      description: Generate and persist a structured catch-up grounded in imported message references
      tags:
        - Chat Export
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                regenerate:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Existing catch-up returned
        '201':
          description: Catch-up generated and persisted
        '404':
          description: Conversation not found for this user
        '413':
          description: Conversation exceeds the bounded summarization limit
        '503':
          description: No real AI provider is configured
  /api/chat-export/extension:
    post:
      summary: Receive extension chat data
      description: Receive chat data from Chrome extension
      tags:
        - Chat Export
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtensionChatData'
      responses:
        '200':
          description: Data received successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtensionChatResponse'
        '400':
          description: Invalid data format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
      required:
        - error
...
    ChatExportResponse:
      type: object
      properties:
        message:
          type: string
        duplicate:
          type: boolean
        data:
          type: object
          properties:
            intakeId:
              type: string
              format: uuid
            displayName:
              type: string
            messageCount:
              type: integer
            participants:
              type: array
              items:
                type: string
            dateRange:
              type: object
              properties:
                start:
                  type: string
                  format: date-time
                end:
                  type: string
                  format: date-time
            dashboardUrl:
              type: string
    ExtensionChatData:
      type: object
      properties:
        chatName:
          type: string
        chatId:
          type: string
        extractedAt:
          type: string
          format: date-time
        messageCount:
          type: integer
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ExtractedMessage'
        source:
          type: string
          enum: [chrome-extension]
        version:
          type: string
        isGroup:
          type: boolean
      required:
        - chatName
        - chatId
        - extractedAt
        - messageCount
        - messages
        - source
        - version
        - isGroup
    ExtractedMessage:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        sender:
          type: string
        timestamp:
          type: string
          format: date-time
        isOutgoing:
          type: boolean
        isMedia:
          type: boolean
        mediaType:
          type: string
          enum: [image, video, audio, document, sticker]
        replyTo:
          type: string
      required:
        - id
        - text
        - sender
        - timestamp
        - isOutgoing
        - isMedia
    ExtensionChatResponse:
      type: object
      properties:
        message:
          type: string
        data:
          type: object
          properties:
            chatId:
              type: string
            intakeId:
              type: string
              format: uuid
            chatName:
              type: string
            messageCount:
              type: integer
            receivedAt:
              type: string
              format: date-time
            dashboardUrl:
              type: string
        duplicate:
          type: boolean
    ConversationSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sourcePlatform:
          type: string
        sourceKind:
          type: string
          enum: [extension, upload]
        displayName:
          type: string
        isGroup:
          type: boolean
        participants:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [received]
        messageCount:
          type: integer
        sourceExtractedAt:
          type: string
          format: date-time
        receivedAt:
          type: string
          format: date-time
    ConversationDetail:
      allOf:
        - $ref: '#/components/schemas/ConversationSummary'
        - type: object
          properties:
            provenance:
              type: object
            messages:
              type: array
              items:
                $ref: '#/components/schemas/ConversationMessage'
    ConversationMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        position:
          type: integer
        senderName:
          type: string
        content:
          type: string
        sentAt:
          type: string
          format: date-time
        isOutgoing:
          type: boolean
        isMedia:
          type: boolean
        mediaType:
          type: string
...